# backend/swagger.yaml
openapi: 3.0.0
info:
  title: RedecellRJ POS API
  version: 1.0.0
  description: API for the RedecellRJ Point of Sale system
servers:
  - url: http://localhost:3000
    description: Development server
tags:
  - name: Authentication
    description: User login and registration
  - name: Products
    description: Product management
  - name: Sales
    description: Sales transactions
  - name: Dashboard
    description: Analytics and reports
  - name: Customers
    description: Customer management
paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                role:
                  type: string
                  enum: ['user', 'admin', 'manager']
                  default: 'user'
      responses:
        '201':
          description: User registered successfully
        '409':
          description: Email already registered
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Log in a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: User logged in successfully
        '401':
          description: Invalid credentials
  /products:
    get:
      tags:
        - Products
      summary: Get all products
      responses:
        '200':
          description: A list of products
    post:
      tags:
        - Products
      summary: Create a new product
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - branch_id
                - variations
              properties:
                name:
                  type: string
                description:
                  type: string
                sku:
                  type: string
                branch_id:
                  type: integer
                variations:
                  type: array
                  items:
                    type: object
                    required:
                      - color
                      - price
                      - stock_quantity
                    properties:
                      color:
                        type: string
                      price:
                        type: number
                      stock_quantity:
                        type: integer
      responses:
        '201':
          description: Product created successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /products/{id}:
    get:
      tags:
        - Products
      summary: Get product by ID
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the product to retrieve
      responses:
        '200':
          description: Product details
        '404':
          description: Product not found
    put:
      tags:
        - Products
      summary: Update product by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the product to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                sku:
                  type: string
                variations:
                  type: array
                  items:
                    type: object
                    properties:
                      id:
                        type: integer
                      color:
                        type: string
                      price:
                        type: number
                      stock_quantity:
                        type: integer
      responses:
        '200':
          description: Product updated successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Product not found
    delete:
      tags:
        - Products
      summary: Delete product by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the product to delete
      responses:
        '204':
          description: Product deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Product not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT